lab5 : some update of lab5, it seems that there is some mistake of func: kernel_execve

+++ labcodes_answer/lab5/kern/syscall/syscall.c
@@ -7,27 +7,27 @@
 #include <assert.h>
 
 static int
-sys_exit(uint32_t arg[]) {
+sys_exit(uint64_t arg[]) {
     int error_code = (int)arg[0];
     return do_exit(error_code);
 }
 
 static int
-sys_fork(uint32_t arg[]) {
+sys_fork(uint64_t arg[]) {
     struct trapframe *tf = current->tf;
     uintptr_t stack = tf->gpr.sp;
     return do_fork(0, stack, tf);
 }
 
 static int
-sys_wait(uint32_t arg[]) {
+sys_wait(uint64_t arg[]) {
     int pid = (int)arg[0];
     int *store = (int *)arg[1];
     return do_wait(pid, store);
 }
 
 static int
-sys_exec(uint32_t arg[]) {
+sys_exec(uint64_t arg[]) {
     const char *name = (const char *)arg[0];
     size_t len = (size_t)arg[1];
     unsigned char *binary = (unsigned char *)arg[2];
@@ -36,35 +36,35 @@ sys_exec(uint32_t arg[]) {
 }
 
 static int
-sys_yield(uint32_t arg[]) {
+sys_yield(uint64_t arg[]) {
     return do_yield();
 }
 
 static int
-sys_kill(uint32_t arg[]) {
+sys_kill(uint64_t arg[]) {
     int pid = (int)arg[0];
     return do_kill(pid);
 }
 
 static int
-sys_getpid(uint32_t arg[]) {
+sys_getpid(uint64_t arg[]) {
     return current->pid;
 }
 
 static int
-sys_putc(uint32_t arg[]) {
+sys_putc(uint64_t arg[]) {
     int c = (int)arg[0];
     cputchar(c);
     return 0;
 }
 
 static int
-sys_pgdir(uint32_t arg[]) {
+sys_pgdir(uint64_t arg[]) {
     print_pgdir();
     return 0;
 }
 
-static int (*syscalls[])(uint32_t arg[]) = {
+static int (*syscalls[])(uint64_t arg[]) = {
     [SYS_exit]              sys_exit,
     [SYS_fork]              sys_fork,
     [SYS_wait]              sys_wait,
@@ -81,7 +81,7 @@ static int (*syscalls[])(uint32_t arg[]) = {
 void
 syscall(void) {
     struct trapframe *tf = current->tf;
-    uint32_t arg[5];
+    uint64_t arg[5];
     int num = tf->gpr.a0;
     if (num >= 0 && num < NUM_SYSCALLS) {
         if (syscalls[num] != NULL) {
